MASSIVE OVERHAUL OF PROGRAM STRUCTURE. Might not be good

+++ source/timer.c
@@ -4,7 +4,7 @@ void start_timer(time_t* p_timer){
     *(p_timer) = time(NULL);
 }
 
-int check_timer(time_t* p_timer, int time_req){
+int check_timer(time_t* p_timer, unsigned int time_req){
     unsigned int time_diff = (unsigned int)(time(NULL) - *(p_timer));
-    return time_diff >= (unsigned int)time_req;
+    return time_diff >= time_req;
 }
